+++ NEWS.md
@@ -1,6 +1,11 @@
 ## v0.1.3
 
 - Function `minimize_vocabulary()` more generic and renamed as `charfilter()`.
+- Method `nchars()` renamed to `charcount()`.
+- Function `duplicates_get_groups()` renamed to `as_docgroups()`.
+- Argument `s_attribute` of method `detect_duplicates()` used generically. A new
+column with the name of the the s-attribute to be used as metadata will be
+added.
 
 ## v0.1.2